api/requirements.txt
autopep8==1.5.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
fire==0.4.0
Flask==1.1.2
Flask-Cors==3.0.10
hurry.filesize==0.9
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.5
pycodestyle==2.6.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
six==1.15.0
termcolor==1.1.0
toml==0.10.2
urllib3==1.26.2
Werkzeug==1.0.1
